backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg[binary]==3.3.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.2.0
httpx>=0.27.0
numpy>=1.26.0

# AI Worker
livekit
livekit-api
ultralytics>=8.2.0
opencv-python-headless

# Behavior Anomaly Detection (SlowFast + MMAction2)
torch>=2.0.0
torchvision>=0.15.0
mmaction2>=1.1.0
mmengine>=0.10.0
mmcv>=2.0.0
